move screens folder to containers
