+++ www/js/containers/App/index.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { Router, Route, Link, IndexRoute, hashHistory, browserHistory } from 'react-router';
 
-import TourScreen from '../../screens/TourScreen';
-import TimelineScreen from '../../screens/TimelineScreen';
-import MapScreen from '../../screens/MapScreen';
-import BrowseScreen from '../../screens/BrowseScreen';
+import TourScreen from '../TourScreen';
+import TimelineScreen from '../TimelineScreen';
+import MapScreen from '../MapScreen';
+import BrowseScreen from '../BrowseScreen';
 
 export default class App extends Component {
   render() {
